+++ gluecodium/src/test/resources/smoke/visibility_attribute/output/android-kotlin/com/example/smoke/InternalClassWithFunctions.kt
@@ -9,11 +9,16 @@ package com.example.smoke
 
 import com.example.NativeBase
 
-class InternalClassWithFunctions : NativeBase {
+internal class InternalClassWithFunctions : NativeBase {
+
+
+
 
     constructor() : this(make(), null as Any?) {
         cacheThisInstance();
     }
+
+
     constructor(foo: String) : this(make(foo), null as Any?) {
         cacheThisInstance();
     }