+++ gluecodium/src/test/resources/smoke/visibility_attribute/output/android-kotlin/com/example/smoke/PublicStructWithInternalConstructors.kt
@@ -13,13 +13,16 @@ class PublicStructWithInternalConstructors {
 
 
 
-    constructor() {
+
+
+   internal constructor() {
         val _other = make()
         this.someVar = _other.someVar
     }
 
 
-    constructor(someVar: Int) {
+
+    internal constructor(someVar: Int) {
         this.someVar = someVar
     }
 
@@ -28,7 +31,8 @@ class PublicStructWithInternalConstructors {
 
 
     companion object {
-        @JvmStatic external fun make() : PublicStructWithInternalConstructors
+
+        @JvmStatic @JvmName("make") internal external fun make() : PublicStructWithInternalConstructors
     }
 }